+++ packages/leemons/src/common/FormWithTheme/components/widgets/BaseInput.js
@@ -1,48 +1,25 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { clone, isDate } from 'lodash';
+import { clone, isDate, isString } from 'lodash';
 import { DatePicker, NumberInput, PasswordInput, TextInput } from '@bubbles-ui/components';
 
-function BaseInput(props) {
-  // Note: since React 15.2.0 we can't forward unknown element attributes, so we
-  // exclude the "options" and "schema" ones here.
+function validateProps(props) {
   if (!props.id) {
     throw new Error(`no id for props ${JSON.stringify(props)}`);
   }
+}
 
-  const {
-    value,
-    readonly,
-    disabled,
-    autofocus,
-    onBlur,
-    onFocus,
-    options,
-    schema,
-    uiSchema,
-    formContext,
-    registry,
-    rawErrors,
-    ...inputProps
-  } = props;
+function getInputProps(props) {
+  const { options, schema, ...inputProps } = props;
 
-  const help = options.help;
-  const description = schema.description;
-
-  // If options.inputType is set use that as the input type
   if (options.inputType) {
     inputProps.type = options.inputType;
   } else if (!inputProps.type) {
-    // If the schema is of type number or integer, set the input type to number
     if (schema.type === 'number') {
       inputProps.type = 'number';
-      // Setting step to 'any' fixes a bug in Safari where decimals are not
-      // allowed in number inputs
       inputProps.step = 'any';
     } else if (schema.type === 'integer') {
       inputProps.type = 'number';
-      // Since this is integer, you always want to step up or down in multiples
-      // of 1
       inputProps.step = '1';
     } else {
       inputProps.type = 'text';
@@ -53,8 +30,6 @@ function BaseInput(props) {
     inputProps.autoComplete = options.autocomplete;
   }
 
-  // If multipleOf is defined, use this as the step value. This mainly improves
-  // the experience for keyboard users (who can use the up/down KB arrows).
   if (schema.multipleOf) {
     inputProps.step = schema.multipleOf;
   }
@@ -67,51 +42,34 @@ function BaseInput(props) {
     inputProps.max = schema.maximum;
   }
 
-  const finalType = clone(inputProps.type);
+  return inputProps;
+}
 
-  const _onChange = (event) => {
-    let newVal;
-    switch (finalType) {
-      case 'number':
-        newVal = event;
-        break;
-      case 'date':
-        newVal = event ? event.toString() : event;
-        break;
-      default:
-        newVal = event;
+function handleDateInputProps(inputProps, schema) {
+  ['minDate', 'maxDate'].forEach((key) => {
+    if (schema[key]) {
+      try {
+        inputProps[key] = new Date(schema[key]);
+      } catch (e) {
+        // ignore
+      }
+      if (!isDate(inputProps[key])) {
+        delete inputProps[key];
+      }
     }
-    return props.onChange({
-      ...value,
-      value: newVal === '' ? options.emptyValue : newVal,
-    });
-  };
+  });
+}
 
+function getInputComponent(inputProps, schema) {
   let Input = null;
-  let finalValue = !value || value.value == null ? '' : value.value;
-
   switch (inputProps.type) {
     case 'number':
       Input = NumberInput;
       break;
     case 'date':
       Input = DatePicker;
-      finalValue = finalValue ? new Date(finalValue) : undefined;
+      handleDateInputProps(inputProps, schema);
       delete inputProps.type;
-      if (schema.minDate || schema.maxDate) {
-        try {
-          inputProps.minDate = new Date(schema.minDate);
-          if (!isDate(inputProps.minDate)) {
-            delete inputProps.minDate;
-          }
-        } catch (e) {}
-        try {
-          inputProps.maxDate = new Date(schema.maxDate);
-          if (!isDate(inputProps.maxDate)) {
-            delete inputProps.maxDate;
-          }
-        } catch (e) {}
-      }
       break;
     case 'password':
       Input = PasswordInput;
@@ -119,6 +77,44 @@ function BaseInput(props) {
       break;
     default:
       Input = TextInput;
+  }
+  return Input;
+}
+
+function handleOnChange(event, finalType, options, value, props) {
+  let newVal;
+  switch (finalType) {
+    case 'number':
+      newVal = event;
+      break;
+    case 'date':
+      newVal = event?.toISOString ? event.toISOString() : event;
+      break;
+    default:
+      newVal = event;
+  }
+  return props.onChange({
+    ...value,
+    value: newVal === '' ? options.emptyValue : newVal,
+  });
+}
+
+function BaseInput(props) {
+  validateProps(props);
+
+  const { value, readonly, disabled, autofocus, onBlur, onFocus, options, schema, rawErrors } =
+    props;
+
+  const inputProps = getInputProps(props);
+  const finalType = clone(inputProps.type);
+
+  const _onChange = (event) => handleOnChange(event, finalType, options, value, props);
+
+  const Input = getInputComponent(inputProps, schema);
+  let finalValue = value?.value ?? '';
+
+  if (inputProps.type === 'date') {
+    finalValue = finalValue ? new Date(finalValue) : undefined;
   }
 
   return [
@@ -128,8 +124,8 @@ function BaseInput(props) {
       readOnly={readonly}
       disabled={disabled}
       autoFocus={autofocus}
-      help={help}
-      description={description}
+      help={options.help}
+      description={schema.description}
       value={finalValue}
       error={rawErrors ? rawErrors[0] : null}
       list={schema.examples ? `examples_${inputProps.id}` : null}
@@ -144,7 +140,7 @@ function BaseInput(props) {
         {[...new Set(schema.examples.concat(schema.default ? [schema.default] : []))].map(
           (example) => (
             <option key={example} value={example} />
-          )
+          ),
         )}
       </datalist>
     ) : null,