+++ packages/leemons/src/common/FormWithTheme/components/widgets/DateTimeWidget.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { utcToLocal, localToUTC } from '../../utils';
 
 function DateTimeWidget(props) {
   const {
@@ -14,7 +15,7 @@ function DateTimeWidget(props) {
       type="datetime-local"
       {...props}
       value={utcToLocal(value)}
-      onChange={(value) => onChange(localToUTC(value))}
+      onChange={(val) => onChange(localToUTC(val))}
     />
   );
 }
@@ -22,8 +23,11 @@ function DateTimeWidget(props) {
 if (process.env.NODE_ENV !== 'production') {
   DateTimeWidget.propTypes = {
     value: PropTypes.string,
+    onChange: PropTypes.func,
+    registry: PropTypes.shape({
+      widgets: PropTypes.object,
+    }),
   };
 }
 
 export default DateTimeWidget;
-import { utcToLocal, localToUTC } from '../../utils';